+++ src/modules/index.ts
@@ -17,4 +17,5 @@ import './interactive/index.js';
 import './approvals/index.js';
 import './scheduling/index.js';
 import './permissions/index.js';
+import './agent-to-agent/index.js';
